+++ src/Symfony/Component/Form/Form.php
@@ -416,13 +416,6 @@ class Form implements \IteratorAggregate, FormInterface, ClearableErrorsInterfac
         return $this;
     }
 
-    public function isSubmittedRequestValid(mixed $request = null): bool
-    {
-        $this->handleRequest($request);
-
-        return $this->isSubmitted() && $this->isValid();
-    }
-
     public function submit(mixed $submittedData, bool $clearMissing = true): static
     {
         if ($this->submitted) {